requirements.txt
accelerate==1.0.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
diffusers==0.31.0
distro==1.9.0
Django==5.0.4
dnspython==2.7.0
entrypoints==0.4
executing==2.0.1
Faker==35.0.0
favicon==0.7.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.55.8
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
graphviz==0.20.3
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
htbuilder==0.9.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.24.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.2
openai==1.53.0
packaging==24.0
pandas==2.2.3
parso==0.8.4
pickle-mixin==1.0.2
pillow==10.4.0
platformdirs==4.2.0
plotly==6.0.0
prometheus_client==0.21.1
prompt-toolkit==3.0.43
proto-plus==1.25.0
protobuf==5.28.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.17.2
pymdown-extensions==10.14.2
pymongo==4.11.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rpds-py==0.20.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.0
st-annotated-text==4.0.2
st-copy-to-clipboard==0.1.6
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.39.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-option-menu==0.4.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
strip-markdown==1.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
transformers==4.46.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
validators==0.34.0
watchdog==5.0.3
wcwidth==0.2.13
Werkzeug==3.0.4
youtube-transcript-api==0.6.2
zipp==3.20.2
numpy>=1.21.0
pandas>=1.3.0
scikit-surprise>=1.1.1
tensorflow>=2.6.0
python-dotenv>=0.19.0